+++ package.json
@@ -29,8 +29,6 @@
     "validator": "^10.11.0"
   },
   "devDependencies": {
-    "eslint": "^7.11.0",
-    "eslint-plugin-react": "^7.21.5",
     "nodemon": "^2.0.4"
   }
 }
